Split different endpoints into their own class;
