+++ Source/Plugins/Core/com.equella.base/src/com/tle/common/settings/standard/SearchSettings.java
@@ -74,7 +74,9 @@ public class SearchSettings implements ConfigurationProperties {
   @Property(key = "search.defaultsort")
   private String defaultSearchSort;
 
-  @JsonIgnore
+  @JsonIgnore // The REST endpoints for this was split into two - one for general search settings,
+              // and one for filters - so we don't want this included in the general search
+              // settings.
   @PropertyDataList(key = SEARCH + ".filters", type = SearchFilter.class)
   private final List<SearchFilter> filters =
       new ArrayList<SearchFilter>() {
