+++ Source/Plugins/Core/com.equella.core/src/com/tle/web/remoting/resteasy/RestEasyServlet.java
@@ -46,6 +46,8 @@ import com.tle.web.api.loginnotice.PostLoginNoticeResource;
 import com.tle.web.api.loginnotice.PreLoginNoticeResource;
 import com.tle.web.api.newuitheme.NewUIThemeResource;
 import com.tle.web.api.searches.SearchConfigApi;
+import com.tle.web.api.settings.CloudSearchSettingsResource;
+import com.tle.web.api.settings.SearchSettingsResource;
 import com.tle.web.api.settings.SettingsResource;
 import com.tle.web.api.users.UserQueryResource;
 import com.tle.web.api.wizard.WizardApi;
@@ -116,6 +118,10 @@ public class RestEasyServlet extends HttpServletDispatcher implements MapperExte
     classes.add(CloudProviderApi.class);
     registry.addSingletonResource(new SettingsResource());
     classes.add(SettingsResource.class);
+    registry.addSingletonResource(new SearchSettingsResource());
+    classes.add(SearchSettingsResource.class);
+    registry.addSingletonResource(new CloudSearchSettingsResource());
+    classes.add(CloudSearchSettingsResource.class);
     registry.addSingletonResource(new LanguageResource());
     classes.add(LanguageResource.class);
     registry.addSingletonResource(new UserQueryResource());
